requirements.txt
python-binance==1.0.19
requests==2.31.0
websocket-client==1.6.4

